requirements.txt
flask
gunicorn
pandas
google-generativeai
python-dotenv